Fixed problem generating submission reference at the beginning of the flow
IVORY-604: Error after share details of item page

The generation of the submission reference has been moved to the beginning of the flow. But it can't be generated on the same page as when the session cookie is created, therefore it has been moved to the next page (/how-certain).
