+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ivory",
-  "version": "0.25.6",
+  "version": "0.25.7",
   "description": "Digital service to support the Ivory Act",
   "main": "index.js",
   "scripts": {
